==> src/2.feature-extraction.ipynb <==
import pandas as pd
import spotipy

# Authentication with Spotify

# Client Flow - to get features and analysis from tracks
# Authenticate my user from Spotify

CLIENT_ID = "CLIENT_ID"
CLIENT_SECRET = "CLIENT_SECRET"
token = spotipy.oauth2.SpotifyClientCredentials(client_id=CLIENT_ID, client_secret=CLIENT_SECRET)
sp = spotipy.Spotify(auth_manager=token)

# Read toptracks_ratings.csv from file 

df_toptracks = pd.read_csv("./dataset/toptracks_ratings.csv")
df_toptracks

# Get track features

def get_features(track_id_list, features_list):
    # get track features from spotify. gets a track id_list and features to extract
    # return a dataframe with track features and id as rows.
    
    # dict and dataframe to store features
    dict_features = {}
    df_features = pd.DataFrame()

    for track_id in track_id_list:
        # make API call and get features from spotify web API
        audio_features = sp.audio_features(track_id)[0]
        # add "id_copy" to double check that the right features are added to the right tracks
        dict_features["id_copy"] = track_id
               
        # add all the wanted features into a dict
        for feature in features_list:
            dict_features[feature] = audio_features[feature]

        # create dataframe of feature dict
        df_track = pd.DataFrame(dict_features, index=[0])
        
        #add to df_features dataframe
        df_features = pd.concat([df_features, df_track], ignore_index=True)
        
    return df_features

# features we want to extract
features_list = [
    "energy", 
    "danceability", 
    "loudness", 
    "valence", 
    "tempo", 
    #"mode", 
    "speechiness", 
    "acousticness", 
    "instrumentalness", 
    "duration_ms", 
    "time_signature", 
    "tempo"
]

# get all track ids from dataset
id_list = list(df_toptracks["id"])

# get features from tracks in a new dataframe
df_features = get_features(id_list, features_list)

df_features.head()

# Add analysis to toptracks dataset

# add more columns to existing data
df_toptracks_features = pd.concat([df_toptracks, df_features], ignore_index=False, axis=1)
df_toptracks_features

# Save to disk

# add to csv
df_toptracks_features.to_csv("./dataset/toptracks_ratings_features.csv", index=False)
