==> src/3.analysis-extraction.ipynb <==
import spotipy
import pandas as pd
import itertools

# Authentication with Spotify

# Client Flow - to get features and analysis from tracks
# Authenticate my user from Spotify

CLIENT_ID = "CLIENT_ID"
CLIENT_SECRET = "CLIENT_SECRET"
token = spotipy.oauth2.SpotifyClientCredentials(client_id=CLIENT_ID, client_secret=CLIENT_SECRET)
sp = spotipy.Spotify(auth_manager=token)

# Load dataset with ratings and features

df_toptracks_features = pd.read_csv("./dataset/toptracks_ratings_features.csv")
df_toptracks_features

# Audio analysis

## define some functions and variables to help us gather and process the analysis data

# a dict specifying which audio analysis features we want.
analysis_features_list = {
    "segments" : # feature parent
    [ # feature array
        "pitches",
        "timbre"
    ],
     "sections" : 
    [
        "key",
        "mode"
    ]
}


def gather_analysis_data(track_id, analysis_features_list, audio_analysis):
    
    # takes track_id, features and analysis object
    # returns a dict where each key is feature names with their values as values.
    
    analysis_data = {} 
    # for every item in the feaures list. val = ["pitches", "timbre"]. Key = "sections", "segments"
    for count1, (feature_parent, feature_array) in enumerate(analysis_features_list.items()):
        
        for feature in feature_array:
            #Initialize a new key
            analysis_data[feature] = []  
            
            # go through the API response and gather all relevant feature values in results dict.
            for count2, (category, analysis) in enumerate(audio_analysis.items()):                
                if category == feature_parent:
                    
                    for item in range(len(analysis)):
                        analysis_data[feature].append(analysis[item][feature])
    
    return analysis_data


def clean_analysis_data(analysis_data, duration_ms):
    
    # Takes analysis_data dict and calculates different kinds of averages of the analysis data lists and numbers.
    # outputs analysis_data dict cleaned, with only 1 value per key
    result = {}
    
    for count, (key, val) in enumerate(analysis_data.items()):       
        averages = []
        
        # if values are timbre or pitches (lists)
        if type(val[0]) == list:
            
            # take the same index items of all lists in the 2d array and average them. [[1,2,3], [1,2,3]] (1+1/2) etc..
            for w in range(len(val[0])):
                temp_store = []
                for item in val:
                    temp_store.append(item[w])
                averages.append(round(sum(temp_store) / len(temp_store), 2))
                
            result[key] = []
            result[key].append(averages)
        
        # if values are scalar (mode, key)
        else:
            if key == "key":
                # take the precentage value indicating how many key changes happen in the track, based on its duration.
                # remove consecutive duplicate values.
                val = [k for k, g in itertools.groupby(val)] 
                
                #how many miliseconds per chord change on average.
                ms_per_chord_change = duration_ms/len(val)
                change_percentage = (100/duration_ms)*ms_per_chord_change
                
                #invert values for a more intuitive reading.
                averages.append(abs(change_percentage-100))
                result[key] = averages
            
            # Take the average mode of the track
            else:
                averages.append(round(sum(val)/len(val), 2))
                result[key] = averages
    
    return result

## Collect, clean, and gather analysis data

df_analysis_result = pd.DataFrame()

# iterate over our toptracks dataframe
for index, row in df_toptracks_features.iterrows():
    #if index < 10:
    analysis_data = {}

    # spotify API call
    audio_analysis = sp.audio_analysis(row["id"])

    # gather raw features
    analysis_data = gather_analysis_data(row["id"], analysis_features_list, audio_analysis)

    # clean data 
    analysis_data_cleaned = clean_analysis_data(analysis_data, row["duration_ms"])

    # concat data into main dict 
    df_analysis_data_cleaned = pd.DataFrame(analysis_data_cleaned)
    df_analysis_data_cleaned["id_copy_2"] = row["id"]
    df_analysis_result = pd.concat([df_analysis_result, df_analysis_data_cleaned], ignore_index=True)
    

# finally, rename some columns for better doc
df_analysis_result = df_analysis_result.rename(columns = {"pitches": "pitch_avg", 
                                                          "timbre" : "timbre_avg", 
                                                          "key": "key_change_percentage", 
                                                          "mode": "mode_avg"}, inplace=False)
# print for check
df_analysis_result

# Add analysis to toptracks dataset

# add more columns to existing data
df_toptracks_features_analysis = pd.concat([df_toptracks_features, df_analysis_result], ignore_index=False, axis=1)
df_toptracks_features_analysis

# these should all be aligned perfectly.
df_check = df_toptracks_features_analysis[["id","id_copy","id_copy_2"]]
df_check

# Save to disk

# add to csv
df_toptracks_features_analysis.to_csv("./dataset/toptracks_ratings_features_analysis.csv", index=False)
