==> src/4.regressor-grid-search.ipynb <==
import ast
import pandas as pd
import numpy as np
import sklearn
import joblib
from sklearn.model_selection import train_test_split
from sklearn.neural_network import MLPRegressor
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import cross_validate
from sklearn.model_selection import RepeatedKFold, GridSearchCV
from sklearn import preprocessing, decomposition, manifold
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

#plotting
import matplotlib
import matplotlib.pyplot as plt

#3d plotting
from mpl_toolkits.mplot3d import Axes3D
#%matplotlib qt
#plt.ion()

# Load dataset

df_toptracks_features_analysis = pd.read_csv("./dataset/toptracks_ratings_features_analysis.csv")
df_toptracks_features_analysis

# Convert dataset to numpy array and clean/prepare for training

# convert dataframe rows into arrays. Every array in the 2d array is a row
features = df_toptracks_features_analysis[
    ["name", "energy", "danceability", "loudness", "valence", "tempo", 
     "speechiness", "acousticness", "instrumentalness", "duration_ms", "time_signature", 
     "pitch_avg", "timbre_avg", "key_change_percentage", "mode_avg"]].to_numpy()

#features = df_toptracks_features_analysis[["name", "pitch_avg", "timbre_avg", "key_change_percentage", "mode_avg"]].to_numpy()

#features = df_toptracks_features_analysis[["name", "energy", "danceability", "loudness", "valence", "tempo", 
#     "speechiness", "acousticness", "instrumentalness", "duration_ms", "time_signature"]].to_numpy()


target = df_toptracks_features_analysis["rating"].to_numpy()
# avoid having float numbers for the training.
# take two decimal points and floor, basically
#target = np.multiply(target, 10)
#target = np.floor(target).astype("int")
target = np.round(target).astype("int")

print(features.shape)
print(features[0])
print(type(features))
print("\n")
print("--------------------")
print(target.shape)
print(target)
print(type(target))

# cleaned features
features_cleaned = features

# Substitute titles with integers
for i in range(features_cleaned.shape[0]):
    features_cleaned[i][0] = i

features_cleaned[0]

# convert the timbre and pitch vectors, which are actually strings in the dataframe imported, to lists again
for row in range(features_cleaned.shape[0]):
    for col in range(features_cleaned.shape[1]):
        if type(features_cleaned[row][col]) == str:
            features_cleaned[row][col] = ast.literal_eval(features_cleaned[row][col])

print(features_cleaned[0])

# unpack/flatten the timbre and pitch lists within the 2d features_cleaned
# (and subesequently extent the coloumn size of the feature array.)

def flatten(x):
    for item in x:
        try:
            #if x has a member (item) it means its a a list or array, therefore we feed the item back into the function.
            yield from flatten(item)
        #so if x has no members to iterate on (i.e its a float or integer), we return it (yield)
        except TypeError:
            yield item

temp_features = np.empty([])
for i in range(features_cleaned.shape[0]):
    
    # flatten row
    row = list(flatten(features_cleaned[i]))
    
    # round all values in row to 2 decimals max
    #row = [round(elem, 2) for elem in row]
    
    # make numpy array of row
    row = np.array(row)
    
    # add them together
    if i == 0:
        temp_features = row
    else:
        temp_features = np.vstack((temp_features, row))

# add to final variable
features_cleaned = temp_features

print(features_cleaned.shape)
print(features_cleaned[0])
print(type(features_cleaned))
print("--------------------")
print(target.shape)
print(target[:10])
print(type(target))

# Dimensionality reduction

# First, we can use this to find how many components I need for the dimentionalty reducction
# to keep 99% of the variance in the dataset

feat_variance = np.var(features_cleaned, axis=0).sum()

for i in range(features_cleaned.shape[1]):
        temp = np.var(features_cleaned[:,0:i+1], axis=0).sum()
        percentage = temp/feat_variance
        if percentage > 0.99:
            print("components needed: ", i+1)
            print("reached: ", percentage, "%")
            break

### First scale the features

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

# learning the scaling
scaler.fit(features_cleaned)

# apply the scaling
features_cleaned = scaler.transform(features_cleaned)


#joblib_file = "scaler.pkl"
#joblib.dump(scaler, joblib_file)
# Hva i helvete skjer her??
#projected_features = sklearn.preprocessing.scale(features_projected)

print(features_cleaned.shape)
features_cleaned[0]

### LDA

components_needed = 2

# lda instance with n-number of components
lda = LinearDiscriminantAnalysis(n_components=components_needed)

# pass/fit the lda model with the features and our target
lda.fit(features_cleaned, target)

features_projected = lda.transform(features_cleaned)

print(features_cleaned.shape)
print(features_projected.shape)
features_projected[0]

# save dimred to file
#joblib_file = "dimred.pkl"
#joblib.dump(lda, joblib_file)

# Grid Search

get a sense of where to start with the tuning of the metrics

mlp = MLPRegressor()
lr = LinearRegression()

print("avaliable model parameters")
print("linear regressor", lr.get_params().keys())
print("MLP", mlp.get_params().keys())

crossval = RepeatedKFold(n_splits=5, n_repeats=20)

params = {
    'activation': ['identity', 'logistic', 'tanh', 'relu'],
    'hidden_layer_sizes': [[100, 50, 30], [200, 100, 50], [100, 75, 50, 25], [100, 100, 50, 50]],
    'max_iter': [1000, 2000, 5000, 10000]
}

gd_sr = GridSearchCV(estimator=mlp,
                     param_grid=params,
                     scoring='r2',
                     cv=crossval,  
                     n_jobs=-1) 

# was features_projected (after dimred)
# features_cleaned (if no dimred)
gd_sr.fit(features_projected, target)

print('best set of parameters', gd_sr.best_params_)
print('associated best score',gd_sr.best_score_)

# try:
# all of this I did with dimred wrong!
# less/other features
# 0 - 5 and 0 - 100 ranges
# try with noizy data and 0-100 0-10
# No dimred 
# PCA instead of LDA

# so far, 0.3 is best r2 score .. with 0-10 range. all features
# with only features, range 0-10, r2 was 0.19
# with all features, range 0-100, r2 was 

## New where i use scaling correctly

* With scale and no dimred. range 0-100. all features. r2 = -0.19
* With scaled and dimred of 10 components, range 0-100, all features. r2 = 0.12
* With scaled and no dimred, range 0-10(!), only analysis features. r2 = 0.13
* With scaled and dimred, 9 components, range 0-10, all features. r2 = 0.3 with {'activation': 'logistic', 'hidden_layer_sizes': [100, 50, 30], 'max_iter': 1000} 
* With scaled and dimred, 2 components, range 0-10, all features. r2 = 0.32 with {'activation': 'tanh', 'hidden_layer_sizes': [100, 50, 30], 'max_iter': 1000}

with a smaller range, there are more datapoints per target.
Apparantly, I should use dimred with target.len -1 components.
So I could keep trying with 10 components.

great link for talking about r2 scores - https://www.statology.org/good-r-squared-value/
