==> src/5.classifier.ipynb <==
import ast
import pandas as pd
import numpy as np
import sklearn

#plotting
import matplotlib
import matplotlib.pyplot as plt

#3d plotting
from mpl_toolkits.mplot3d import Axes3D
#%matplotlib qt
#plt.ion()

# Load and clean features

df_toptracks_features_analysis = pd.read_csv("./dataset/toptracks_ratings_features_analysis.csv")
df_toptracks_features_analysis

# convert dataframe rows into arrays. Every array in the 2d array is a row
features = df_toptracks_features_analysis[
    ["name", "energy", "danceability", "loudness", "valence", "tempo", 
     "speechiness", "acousticness", "instrumentalness", "duration_ms", "time_signature", 
    "pitch_avg", "timbre_avg", "key_change_percentage", "mode_avg"]].to_numpy()

# my 5 classes
classes = np.array(["bad", "fine", "good", "great"])
numb_classes = len(classes)

# convert all ratings from 0 - 10 into 0 - 4, to match the classes
labels = df_toptracks_features_analysis["rating"].to_numpy()
labels = np.round([label/10*(numb_classes-1) for label in labels]).astype("int")

print("first row of features:", features[0])
print(features.shape)
print("\n")
print("my classes:", classes)
print(classes.shape)
print("\n")
print("my labels:", labels)
print(labels.shape)

features_cleaned = features

# Substitute titles with integers
for i in range(features_cleaned.shape[0]):
    features_cleaned[i][0] = i

features_cleaned[0]

# convert the timbre and pitch vectors, which are actually strings in the dataframe imported, to lists again
for row in range(features_cleaned.shape[0]):
    for col in range(features_cleaned.shape[1]):
        if type(features_cleaned[row][col]) == str:
            features_cleaned[row][col] = ast.literal_eval(features_cleaned[row][col])

print(features_cleaned[0])

# unpack/flatten the timbre and pitch lists within the 2d features_cleaned
# (and subesequently extent the coloumn size of the feature array.)

def flatten(x):
    for item in x:
        try:
            #if x has a member (item) it means its a a list or array, therefore we feed the item back into the function.
            yield from flatten(item)
        #so if x has no members to iterate on (i.e its a float or integer), we return it (yield)
        except TypeError:
            yield item

temp_features = np.empty([])
for i in range(features_cleaned.shape[0]):
    
    # flatten row
    row = list(flatten(features_cleaned[i]))
    
    # round all values in row to 2 decimals max
    #row = [round(elem, 2) for elem in row]
    
    # make numpy array of row
    row = np.array(row)
    
    # add them together
    if i == 0:
        temp_features = row
    else:
        temp_features = np.vstack((temp_features, row))

# add to final variable
features_cleaned = temp_features

print(features_cleaned.shape)
print(features_cleaned[0])

# Create pipeline
The best result of the grid search was:

* LDA, 5 classes, all features, kfold(5, 100). {'activation': 'identity', 'hidden_layer_sizes': [100, 50, 30], 'max_iter': 1000}. Accuracy = 0.58
* LDA, 4 classes, all features, kfold(5,100). {'activation': 'identity', 'hidden_layer_sizes': [100, 100, 50, 50], 'max_iter': 500}


from sklearn.pipeline import Pipeline
from sklearn.neural_network import MLPClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

mlp = MLPClassifier(hidden_layer_sizes=([100, 100, 50, 50]), max_iter=1000, activation='identity')

svm = sklearn.svm.SVC(kernel='linear', C=1.0)

#creating pipeline
pipe = Pipeline([
        ('scaler', StandardScaler()),
        ('dim_red', LinearDiscriminantAnalysis()),
        ('classifier', mlp)
        ])

# Scale the features
#from sklearn.preprocessing import StandardScaler
#from sklearn.model_seblection import train_test_split

#feat_train, feat_test, lab_train, lab_test = train_test_split(features_cleaned, labels, test_size=0.2, random_state=10)

#creating scaling object
#scaler = sklearn.preprocessing.StandardScaler()

#learning scaling from train set
#scaler.fit(feat_train)

#applying scaling to both train and test set
#feat_train = scaler.transform(feat_train)
#feat_test = scaler.transform(feat_test)

#displaying the size of the original and extended arrays
#print(feat_train.shape)
#print(lab_train.shape)

from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.model_selection import RepeatedKFold

rkf = RepeatedStratifiedKFold(n_splits=5, n_repeats=100)

#initializing the cross validator with pipe, features, target, scores, and kfold object
scores = sklearn.model_selection.cross_validate(pipe, features_cleaned, labels, cv=rkf, scoring=('f1_macro', 'accuracy'), return_train_score=True)

#pipe.fit(feat_train, lab_train)
#lab_predict = pipe.predict(feat_test)

print(scores,'\n')
print('Accuracy mean and variance', np.mean(scores['test_accuracy']),np.var(scores['test_accuracy']),'\n')
print('F1 macro mean and variance', np.mean(scores['test_f1_macro']),np.var(scores['test_f1_macro']),'\n')

#plt.plot(mlp.loss_curve_)
#plt.xlabel('iteration')
#plt.ylabel('validation loss')
#plt.title("MLPClassifier loss curve")
#plt.show()

#print the number of misclassified samples, accuracy and complete report (using scikit learn metric tools) 
#print('Number of mislabeled samples %d out of %d' % ((lab_test != lab_predict).sum(),lab_test.size))
#print('Accuracy:',sklearn.metrics.accuracy_score(lab_test, lab_predict))
#print(sklearn.metrics.classification_report(lab_test, lab_predict))
#print('confusion matrix')
#print(sklearn.metrics.confusion_matrix(lab_test,lab_predict))

# Plotting the train and validation learning curves

If i take this as a reference: 
https://towardsdatascience.com/learning-curve-to-identify-overfitting-underfitting-problems-133177f38df5,
https://www.datacamp.com/tutorial/tutorial-learning-curves

the drop of traing validation from high to lower can be interperted as an underfitting model. This phenonemon indicates that the addition of more training examples can’t improve the model performance on unseen data.

Learning curves show the effect of adding more samples during the training process.

The test set is used to evaluate the performance of your model.


"My model starts with high accuracy, but then it drops (it starts gaining more loss) as we show it more and more data. Over time. in other words, the addition of more training examples can’t improve the model performance on unseen data." 

from sklearn.model_selection import learning_curve
from sklearn.model_selection import RepeatedKFold
from sklearn.datasets import load_digits
from sklearn.neural_network import MLPClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

mlp = MLPClassifier(hidden_layer_sizes=([100, 100, 50, 50]), max_iter=500, activation='identity')

#creating scaling object
scaler = sklearn.preprocessing.StandardScaler()
scaler.fit(features_cleaned)
features_cleaned = scaler.transform(features_cleaned)

# dimentionality reduction with LDA (down to 3 features)
lda = LinearDiscriminantAnalysis()
lda.fit(features_cleaned, labels)
features_projected = lda.transform(features_cleaned)

crossval = RepeatedKFold(n_splits=5, n_repeats=100)
train_sizes, train_scores, test_scores, fit_times, _ = learning_curve(mlp, 
                                                                      features_projected, 
                                                                      labels, 
                                                                      cv=crossval, 
                                                                      scoring="accuracy",
                                                                      return_times=True)

train_scores = 1-np.mean(train_scores,axis=1)
test_scores = 1-np.mean(test_scores,axis=1)

#converting the accuracy score to misclassification rate/loss
#train_scores = 1-np.mean(train_scores,axis=1)
#test_scores = 1-np.mean(test_scores,axis=1)

# plot
plt.plot(train_sizes, train_scores)
plt.plot(train_sizes, test_scores)
plt.xlabel('dataset size')
#plt.ylabel('misclassification rate/loss')
plt.ylabel('accuracy (score)')
plt.title("MLPClassifier learning curve")
plt.legend(["training", "validation"])
plt.show()
